# Conflicts:
#	src/main/java/so/sao/shop/supplier/config/Swagger2Configure.java

+++ src/main/java/so/sao/shop/supplier/web/AccountController.java
@@ -1,12 +1,13 @@
 package so.sao.shop.supplier.web;
 
-import com.aliyuncs.exceptions.ClientException;
 import com.github.pagehelper.PageInfo;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.AuthenticationException;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.ObjectError;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -18,12 +19,13 @@ import so.sao.shop.supplier.service.*;
 import so.sao.shop.supplier.util.Constant;
 import so.sao.shop.supplier.util.ExcelImportUtils;
 
-import javax.servlet.ServletOutputStream;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.validation.Valid;
 import java.io.*;
-import java.net.URL;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * Created by xujc on 2017/7/18.
@@ -53,30 +55,29 @@ public class AccountController {
      *
      * @param account 供应商信息
      * @return 返回添加结果
-     * @throws IOException
-     * @throws ClientException
      */
     @ApiOperation("添加供应商信息")
     @PostMapping("/save")
-    public BaseResult save(@RequestBody Account account) {
+    public BaseResult save(HttpServletRequest request, @Valid @RequestBody Account account,BindingResult result) {
         BaseResult baseResult = new BaseResult();
-        /**
-         * 插入用户信息
-         */
-        Long id = accountService.saveUser(account.getContractResponsiblePhone());
-        if (id != 0l) {
-            account.setUserId(id);
-        } else {
-            baseResult.setMessage("插入用户信息失败!");
+        User user = (User) request.getAttribute(Constant.REQUEST_USER);
+        if(user==null || !user.getIsAdmin().equals(Constant.IS_ADMIN)){
+            baseResult.setCode(0);
+            baseResult.setMessage(so.sao.shop.supplier.config.Constant.MessageConfig.ADMIN_AUTHORITY_EERO);
             return baseResult;
         }
-        int num = accountService.insert(account);
-        if (num < 0) {
-            baseResult.setCode(0);
-            baseResult.setMessage("添加失败");
-        } else {
-            baseResult.setCode(1);
-            baseResult.setMessage("添加成功");
+        //判断验证是否通过。true 未通过  false通过
+        if(result.hasErrors()) {
+            List<ObjectError> list = result.getAllErrors();
+            for (ObjectError error : list) {
+                baseResult.setCode(so.sao.shop.supplier.config.Constant.CodeConfig.CODE_NOT_EMPTY);
+                baseResult.setMessage(error.getDefaultMessage());
+            }
+        }else{
+            /**
+             * 插入用户和供应商信息
+             */
+            return accountService.saveUserAndAccount(account);
         }
         return baseResult;
     }
@@ -99,21 +100,34 @@ public class AccountController {
      * @return 返回更新成功或者失败
      */
     @ApiOperation("修改供应商信息")
-    @PostMapping("/update")
-    public BaseResult update(@RequestBody Account account) {
-        //修改用户登录名
-        User user = new User();
-        user.setUsername(account.getResponsiblePhone());
-        accountService.updateUser(account.getUserId());
-        // 修改用户信息
-        int num = accountService.update(account);
+    @PutMapping("/update")
+    public BaseResult update(@Valid @RequestBody Account account,BindingResult result, HttpServletRequest request) {
         BaseResult baseResult = new BaseResult();
-        if (num < 0) {
+        User user = (User) request.getAttribute(Constant.REQUEST_USER);
+        if(user==null || !user.getIsAdmin().equals(Constant.IS_ADMIN)){
             baseResult.setCode(0);
-            baseResult.setMessage("修改失败");
-        } else {
-            baseResult.setCode(1);
-            baseResult.setMessage("修改成功");
+            baseResult.setMessage(so.sao.shop.supplier.config.Constant.MessageConfig.ADMIN_AUTHORITY_EERO);
+            return baseResult;
+        }
+        //判断验证是否通过。true 未通过  false通过
+        if(result.hasErrors()) {
+            List<ObjectError> list = result.getAllErrors();
+            for (ObjectError error : list) {
+                baseResult.setCode(0);
+                baseResult.setMessage(error.getDefaultMessage());
+            }
+        }else{
+            //修改用户登录名
+            accountService.updateUser(account.getUserId(), account.getContractResponsiblePhone());
+            // 修改用户信息
+            int num = accountService.update(account);
+            if (num < 0) {
+                baseResult.setCode(0);
+                baseResult.setMessage("修改失败");
+            } else {
+                baseResult.setCode(1);
+                baseResult.setMessage("修改成功");
+            }
         }
         return baseResult;
     }
@@ -125,10 +139,16 @@ public class AccountController {
      * @return 返回成功或失败
      */
     @ApiOperation("删除供应商信息")
-    @GetMapping("/delete/{id}")
-    public BaseResult delete(@PathVariable Long id) {
-        int num = accountService.delete(id);
+    @DeleteMapping("/delete/{id}")
+    public BaseResult delete(@PathVariable Long id, HttpServletRequest request) {
         BaseResult baseResult = new BaseResult();
+        User user = (User) request.getAttribute(Constant.REQUEST_USER);
+        if(user==null || !user.getIsAdmin().equals(Constant.IS_ADMIN)){
+            baseResult.setCode(0);
+            baseResult.setMessage(so.sao.shop.supplier.config.Constant.MessageConfig.ADMIN_AUTHORITY_EERO);
+            return baseResult;
+        }
+        int num = accountService.delete(id);
         if (num < 0) {
             baseResult.setCode(0);
             baseResult.setMessage("删除失败");
@@ -145,7 +165,7 @@ public class AccountController {
      * @return
      */
     @ApiOperation("银行信息")
-    @PostMapping("/selectBank")
+    @GetMapping("/selectBank")
     public List<String> selectBank() {
         return accountService.selectBank();
     }
@@ -156,9 +176,20 @@ public class AccountController {
      * @return
      */
     @ApiOperation("行业信息")
-    @PostMapping("/selectHangYe")
+    @GetMapping("/selectHangYe")
     public List<String> selectHangYe() {
         return accountService.selectHangYe();
+    }
+
+    /**
+     * 初始化行业信息
+     *
+     * @return
+     */
+    @ApiOperation("行业信息(字典项)")
+    @GetMapping("/selectHangYeDict")
+    public List<DictItem> selectHangYeDict() {
+        return accountService.selectHangYeDict();
     }
 
     /**
@@ -174,8 +205,48 @@ public class AccountController {
     public Result<String> createAuthenticationToken(
             User authenticationRequest, HttpServletResponse response) throws AuthenticationException, IOException {
 
-        final String token = authService.login(authenticationRequest.getUsername(), authenticationRequest.getPassword());
-        return new Result<String>(1, "", token);
+        return authService.login(authenticationRequest.getUsername(), authenticationRequest.getPassword());
+    }
+
+    /**
+     * 获取admin标记(一期暂时)
+     * @param request
+     * @return
+     */
+    @ApiOperation("获取admin标记")
+    @GetMapping(value = "/isAdmin")
+    public Result isAdmin(HttpServletRequest request){
+        Map result = new HashMap();
+        User user = (User) request.getAttribute(Constant.REQUEST_USER);
+        result.put("isAdmin",user.getIsAdmin());
+        return new Result(1,"",result);
+    }
+
+    /**
+     * 获取当前登陆供应商名
+     * @param request
+     * @return
+     */
+    @ApiOperation("获取当前登陆供应商名")
+    @GetMapping(value = "/username")
+    public Result getUserName(HttpServletRequest request){
+        Map result = new HashMap();
+        User user = (User) request.getAttribute(Constant.REQUEST_USER);
+        Account account = accountService.selectByUserId(user.getId());
+        result.put("username",account!=null?account.getProviderName():"");
+        return new Result(1,"",result);
+    }
+
+    /**
+     * 登出,设置登出时间
+     *
+     * @return
+     */
+    @ApiOperation("登出")
+    @PostMapping(value = "/logout")
+    public Result<String> logout(HttpServletRequest request) {
+        User user = (User) request.getAttribute(Constant.REQUEST_USER);
+        return authService.logout(user.getId());
     }
 
     /**
@@ -195,12 +266,11 @@ public class AccountController {
      *
      * @param tel
      * @return
-     * @throws AuthenticationException
      * @throws IOException
      */
     @ApiOperation("忘记密码")
     @GetMapping(value = "/findPassword/{tel}")
-    public BaseResult getPassword(@PathVariable String tel) throws IOException, ClientException {
+    public BaseResult getPassword(HttpServletRequest request, @PathVariable String tel) throws IOException {
         return authService.getPassword(tel);
     }
 
@@ -212,7 +282,11 @@ public class AccountController {
      */
     @GetMapping(value = "/account")
     @ApiOperation(value = "查询供应商列表")
-    public PageInfo search(Condition condition) {
+    public PageInfo search(Condition condition, HttpServletRequest request)  {
+        User user = (User) request.getAttribute(Constant.REQUEST_USER);
+        if(user==null || !user.getIsAdmin().equals(Constant.IS_ADMIN)){
+            throw new RuntimeException("unauthorized access");
+        }
         return accountService.searchAccount(condition);
     }
 
@@ -224,9 +298,41 @@ public class AccountController {
      */
 
     @GetMapping("/search")
-    @ApiOperation(value = "根据id查询供应商")
-    public Account get(@RequestParam Long id) {
+    @ApiOperation(value = "根据id查询供应商(省市区汉字)")
+    public Account get(@RequestParam(required = false) Long id, HttpServletRequest request){
+        id = getAccountId(id, request);
         return accountService.selectById(id);
+    }
+
+    /**
+     * 根据id查询供应商详细信息
+     *
+     * @param id
+     * @return
+     */
+    @GetMapping("/search0")
+    @ApiOperation(value = "根据id查询供应商(省市区编码)")
+    public Account get0(@RequestParam(required = false) Long id, HttpServletRequest request) {
+        id = getAccountId(id, request);
+        return accountService.selectById0(id);
+    }
+
+    /**
+     * 根据当前登录人角色获取accountId
+     * @param id
+     * @param request
+     * @return
+     */
+    private Long getAccountId(Long id, HttpServletRequest request){
+        User user = (User) request.getAttribute(Constant.REQUEST_USER);
+        if(user==null){
+            throw new RuntimeException("unauthorized access");
+        }else if(user.getIsAdmin().equals(Constant.IS_ADMIN )&&(id == null || id == 0)){
+            throw new RuntimeException("unauthorized access");
+        }else if(!user.getIsAdmin().equals(Constant.IS_ADMIN )){
+            id = user.getAccountId();
+        }
+        return id;
     }
 
     /**
@@ -249,11 +355,10 @@ public class AccountController {
      * @param tel
      * @return
      * @throws IOException
-     * @throws ClientException
      */
     @ApiOperation("发送验证码")
     @GetMapping(value = "/sendCode/{tel}")
-    public BaseResult sendCode(@PathVariable String tel) throws IOException, ClientException {
+    public BaseResult sendCode(@PathVariable String tel) throws IOException {
         return authService.sendCode(tel);
     }
 
@@ -265,7 +370,7 @@ public class AccountController {
      */
     @ApiOperation("检验验证码")
     @GetMapping(value = "/verifySmsCode/{code}")
-    public BaseResult verifySmsCode(HttpServletRequest request, String code) {
+    public BaseResult verifySmsCode(HttpServletRequest request, @PathVariable String code) {
         User user = (User) request.getAttribute(Constant.REQUEST_USER);
         return authService.verifySmsCode(user.getId(), code);
     }
@@ -278,57 +383,10 @@ public class AccountController {
      * @return
      */
     @ApiOperation("密码修改")
-    @PutMapping(value = "/updatePassword")
-    public BaseResult updatePassword(HttpServletRequest request, String encodedPassword) {
+    @PutMapping(value = "/updatePassword/{encodedPassword}")
+    public BaseResult updatePassword(HttpServletRequest request, @PathVariable String encodedPassword) {
         User user = (User) request.getAttribute(Constant.REQUEST_USER);
         return authService.updatePassword(user.getId(), encodedPassword);
-    }
-
-    /**
-     * 供应商信息模板下载
-     *
-     * @param request
-     * @param response
-     * @throws IOException
-     */
-    @ApiOperation("供应商信息模板下载")
-    @GetMapping("/down")
-    public void downLoadExcel(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        URL save = Thread.currentThread().getContextClassLoader().getResource("");
-        String str = save.toString();
-        str = str.substring(6, str.length());
-        str = str.replaceAll("%20", " ");
-        int num = str.lastIndexOf("supplier");//supplier 为项目名，应用到不同的项目中，这个需要修改！
-        str = str.substring(0, num + "supplier".length());
-        str = str + "/file/template.xlsx";//Excel模板所在的路径。
-        File f = new File(str);
-        // 设置response参数，可以打开下载页面
-        response.reset();
-        response.setContentType("application/vnd.ms-excel;charset=utf-8");
-        String filename = "供应商信息.xlsx";
-        filename = new String(filename.getBytes("Utf-8"), "iso-8859-1");
-        response.setHeader("Content-Disposition", "attachment; filename=\"" + filename + "\"");
-        response.setCharacterEncoding("UTF-8");
-        response.setHeader("Content-Type", "application/octet-stream");
-        ServletOutputStream out = response.getOutputStream();
-        BufferedInputStream bis = null;
-        BufferedOutputStream bos = null;
-        try {
-            bis = new BufferedInputStream(new FileInputStream(f));
-            bos = new BufferedOutputStream(out);
-            byte[] buff = new byte[2048];
-            int bytesRead;
-            while (-1 != (bytesRead = bis.read(buff, 0, buff.length))) {
-                bos.write(buff, 0, bytesRead);
-            }
-        } catch (final IOException e) {
-            throw e;
-        } finally {
-            if (bis != null)
-                bis.close();
-            if (bos != null)
-                bos.close();
-        }
     }
 
     /**
@@ -339,8 +397,8 @@ public class AccountController {
      */
 
     @ApiOperation("供应商上传记录")
-    @PostMapping("/record")
-    public PageInfo<SupplierRecord> findAccount(@RequestBody Condition condition) {
+    @GetMapping("/record")
+    public PageInfo<SupplierRecord> findAccount(Condition condition) {
         return supplierRecordService.searchAccountRecord(condition);
     }
 
@@ -361,7 +419,11 @@ public class AccountController {
      */
     @ApiOperation("供应商信息上传")
     @PostMapping("/upload")
-    public String excelUpload(HttpServletRequest request, HttpServletResponse response, @RequestBody MultipartFile file) {
+    public String excelUpload(HttpServletRequest request, HttpServletResponse response, MultipartFile file) {
+        User user = (User) request.getAttribute(Constant.REQUEST_USER);
+        if(user==null || !user.getIsAdmin().equals(Constant.IS_ADMIN)){
+            throw new RuntimeException("unauthorized access");
+        }
         //判断文件是否为空
         if (file == null) {
             return "文件不能为空";
@@ -377,7 +439,7 @@ public class AccountController {
 
         //进一步判断文件内容是否为空（即判断其大小是否为0或其名称是否为null）
         long size = file.getSize();
-        if (StringUtils.isEmpty(fileName) || size == 0) {
+        if (StringUtils.isEmpty(fileName) || size <= 9700) {
             return "文件内容为空";
         }
 
@@ -385,5 +447,4 @@ public class AccountController {
         String message = importExcel.batchImport(fileName, file);
         return message;
     }
-
 }
